fix: make sure no imports from @stitches/react alone
